+++ gateway-js/src/__tests__/gateway/reporting.test.ts
@@ -228,6 +228,7 @@ describe('reporting', () => {
           "seconds": "1562203363",
         },
         "header": "<HEADER>",
+        "operationCount": 1,
         "tracesPerQuery": Object {
           "# -
       {me{name{first last}}topProducts{name}}": Object {
@@ -299,6 +300,7 @@ describe('reporting', () => {
                               "nanos": 123000000,
                               "seconds": "1562203363",
                             },
+                            "fieldExecutionWeight": 1,
                             "root": Object {
                               "child": Array [
                                 Object {
@@ -364,6 +366,7 @@ describe('reporting', () => {
                                     "nanos": 123000000,
                                     "seconds": "1562203363",
                                   },
+                                  "fieldExecutionWeight": 1,
                                   "root": Object {
                                     "child": Array [
                                       Object {
@@ -465,6 +468,7 @@ describe('reporting', () => {
                                         "nanos": 123000000,
                                         "seconds": "1562203363",
                                       },
+                                      "fieldExecutionWeight": 1,
                                       "root": Object {
                                         "child": Array [
                                           Object {
@@ -568,6 +572,7 @@ describe('reporting', () => {
                                         "nanos": 123000000,
                                         "seconds": "1562203363",
                                       },
+                                      "fieldExecutionWeight": 1,
                                       "root": Object {
                                         "child": Array [
                                           Object {